#346 create brick table models
