+++ src/f09_brick/test_brick_config/test_brick_config_.py
@@ -1,6 +1,6 @@
 from src.f00_instrument.dict_tool import get_json_from_dict
 from src.f00_instrument.file import save_file
-from src.f01_road.finance_tran import time_id_str, road_delimiter_str
+from src.f01_road.finance_tran import quota_str, time_id_str, road_delimiter_str
 from src.f02_bud.bud_tool import (
     bud_acct_membership_str,
     bud_acctunit_str,
@@ -20,22 +20,57 @@ from src.f03_chrono.chrono import (
     yr1_jan1_offset_str,
 )
 from src.f04_gift.atom_config import (
-    face_id_str,
-    penny_str,
-    respect_bit_str,
-    fund_coin_str,
+    get_atom_args_category_mapping,
     get_atom_config_dict,
+    get_atom_categorys,
+    required_args_str,
+    optional_args_str,
+    normal_specs_str,
+    column_order_str,
     atom_delete,
     atom_insert,
     atom_update,
-    normal_specs_str,
-    required_args_str,
-    optional_args_str,
     face_id_str,
+    obj_class_str,
     fiscal_id_str,
-    get_atom_categorys,
+    owner_id_str,
+    acct_id_str,
+    group_id_str,
+    parent_road_str,
+    label_str,
+    road_str,
+    base_str,
+    team_id_str,
+    awardee_id_str,
+    healer_id_str,
+    numor_str,
+    denom_str,
+    addin_str,
+    base_item_active_requisite_str,
+    begin_str,
+    close_str,
+    credit_belief_str,
+    debtit_belief_str,
+    credit_vote_str,
+    debtit_vote_str,
+    credor_respect_str,
+    debtor_respect_str,
+    fopen_str,
+    fnigh_str,
+    gogo_want_str,
+    mass_str,
+    morph_str,
+    pledge_str,
+    stop_want_str,
+    fund_coin_str,
+    penny_str,
+    respect_bit_str,
+    give_force_str,
+    take_force_str,
 )
+
 from src.f07_fiscal.fiscal_config import (
+    get_fiscal_args_category_mapping,
     get_fiscal_config_dict,
     get_fiscal_categorys,
     fiscalunit_str,
@@ -46,18 +81,35 @@ from src.f07_fiscal.fiscal_config import (
     fiscal_timeline_month_str,
     fiscal_timeline_weekday_str,
     current_time_str,
+    amount_str,
+    month_label_str,
+    hour_label_str,
+    cumlative_minute_str,
+    cumlative_day_str,
+    weekday_label_str,
+    weekday_order_str,
 )
 from src.f08_filter.filter_config import (
     eon_id_str,
     filterunit_str,
+    otx_road_delimiter_str,
+    inx_road_delimiter_str,
+    unknown_word_str,
+    otx_word_str,
+    inx_word_str,
+    otx_label_str,
+    inx_label_str,
     bridge_explicit_label_str,
     bridge_otx_to_inx_str,
     get_filter_categorys,
     get_filter_config_dict,
+    get_filter_args_category_mapping,
 )
 from src.f09_brick.brick_config import (
     brick_type_str,
     get_brick_types,
+    get_brick_elements_sort_order,
+    get_brick_sqlite_type,
     brick_number_str,
     allowed_crud_str,
     attributes_str,
@@ -81,7 +133,6 @@ from src.f09_brick.brick_config import (
     brick_format_00020_bud_acct_membership_v0_0_0,
     brick_format_00013_itemunit_v0_0_0,
 )
-from src.f08_filter.filter_config import get_filter_categorys
 from os import getcwd as os_getcwd
 
 
@@ -101,6 +152,177 @@ def test_str_functions_ReturnObj():
     assert delete_update_str() == "DELETE_UPDATE"
 
     assert get_brick_types() == {budunit_str(), fiscalunit_str(), filterunit_str()}
+
+
+def test_get_brick_elements_sort_order_ReturnsObj():
+    # ESTABLISH / WHEN
+    table_sorting_priority = get_brick_elements_sort_order()
+
+    # THEN
+    assert table_sorting_priority[0] == face_id_str()
+    assert table_sorting_priority[1] == eon_id_str()
+    assert table_sorting_priority[2] == fiscal_id_str()
+    assert table_sorting_priority[3] == obj_class_str()
+    assert table_sorting_priority[4] == owner_id_str()
+    assert table_sorting_priority[5] == acct_id_str()
+    assert table_sorting_priority[6] == group_id_str()
+    assert table_sorting_priority[7] == parent_road_str()
+    assert table_sorting_priority[8] == label_str()
+    assert table_sorting_priority[9] == road_str()
+    assert table_sorting_priority[10] == base_str()
+    assert table_sorting_priority[11] == "need"
+    assert table_sorting_priority[12] == "pick"
+    assert table_sorting_priority[13] == team_id_str()
+    assert table_sorting_priority[14] == awardee_id_str()
+    assert table_sorting_priority[15] == healer_id_str()
+    assert table_sorting_priority[16] == time_id_str()
+    assert table_sorting_priority[17] == begin_str()
+    assert table_sorting_priority[18] == close_str()
+    assert table_sorting_priority[19] == addin_str()
+    assert table_sorting_priority[20] == numor_str()
+    assert table_sorting_priority[21] == denom_str()
+    assert table_sorting_priority[22] == morph_str()
+    assert table_sorting_priority[23] == gogo_want_str()
+    assert table_sorting_priority[24] == stop_want_str()
+    assert table_sorting_priority[25] == base_item_active_requisite_str()
+    assert table_sorting_priority[26] == credit_belief_str()
+    assert table_sorting_priority[27] == debtit_belief_str()
+    assert table_sorting_priority[28] == credit_vote_str()
+    assert table_sorting_priority[29] == debtit_vote_str()
+    assert table_sorting_priority[30] == credor_respect_str()
+    assert table_sorting_priority[31] == debtor_respect_str()
+    assert table_sorting_priority[32] == fopen_str()
+    assert table_sorting_priority[33] == fnigh_str()
+    assert table_sorting_priority[34] == "fund_pool"
+    assert table_sorting_priority[35] == give_force_str()
+    assert table_sorting_priority[36] == mass_str()
+    assert table_sorting_priority[37] == "max_tree_traverse"
+    assert table_sorting_priority[38] == "nigh"
+    assert table_sorting_priority[39] == "open"
+    assert table_sorting_priority[40] == "divisor"
+    assert table_sorting_priority[41] == pledge_str()
+    assert table_sorting_priority[42] == "problem_bool"
+    assert table_sorting_priority[43] == "purview_time_id"
+    assert table_sorting_priority[44] == take_force_str()
+    assert table_sorting_priority[45] == "tally"
+    assert table_sorting_priority[46] == fund_coin_str()
+    assert table_sorting_priority[47] == penny_str()
+    assert table_sorting_priority[48] == respect_bit_str()
+    assert table_sorting_priority[49] == current_time_str()
+    assert table_sorting_priority[50] == amount_str()
+    assert table_sorting_priority[51] == month_label_str()
+    assert table_sorting_priority[52] == hour_label_str()
+    assert table_sorting_priority[53] == cumlative_minute_str()
+    assert table_sorting_priority[54] == cumlative_day_str()
+    assert table_sorting_priority[55] == weekday_label_str()
+    assert table_sorting_priority[56] == weekday_order_str()
+    assert table_sorting_priority[57] == otx_road_delimiter_str()
+    assert table_sorting_priority[58] == inx_road_delimiter_str()
+    assert table_sorting_priority[59] == unknown_word_str()
+    assert table_sorting_priority[60] == otx_word_str()
+    assert table_sorting_priority[61] == inx_word_str()
+    assert table_sorting_priority[62] == otx_label_str()
+    assert table_sorting_priority[63] == inx_label_str()
+    assert table_sorting_priority[64] == road_delimiter_str()
+    assert table_sorting_priority[65] == c400_number_str()
+    assert table_sorting_priority[66] == yr1_jan1_offset_str()
+    assert table_sorting_priority[67] == quota_str()
+    assert table_sorting_priority[68] == monthday_distortion_str()
+    assert table_sorting_priority[69] == timeline_label_str()
+    assert len(table_sorting_priority) == 70
+    atom_args = set(get_atom_args_category_mapping().keys())
+    assert atom_args.issubset(set(table_sorting_priority))
+    fiscal_args = set(get_fiscal_args_category_mapping().keys())
+    print(f"{fiscal_args=}")
+    print(f"{fiscal_args.difference(set(table_sorting_priority))=}")
+    assert fiscal_args.issubset(set(table_sorting_priority))
+    filter_args = set(get_filter_args_category_mapping().keys())
+    assert filter_args.issubset(set(table_sorting_priority))
+    atom_fiscal_filter_args = atom_args
+    atom_fiscal_filter_args.update(fiscal_args)
+    atom_fiscal_filter_args.update(filter_args)
+    table_sorting_priority.remove(eon_id_str())
+    table_sorting_priority.remove(face_id_str())
+    table_sorting_priority.remove(obj_class_str())
+    assert atom_fiscal_filter_args == set(table_sorting_priority)
+
+
+def test_get_brick_sqlite_type_ReturnsObj():
+    # ESTABLISH / WHEN
+    sqlite_types = get_brick_sqlite_type()
+
+    # THEN
+    assert set(sqlite_types.keys()) == set(get_brick_elements_sort_order())
+    assert sqlite_types.get(face_id_str()) == "TEXT"
+    assert sqlite_types.get(eon_id_str()) == "INTEGER"
+    assert sqlite_types.get(fiscal_id_str()) == "TEXT"
+    assert sqlite_types.get(obj_class_str()) == "TEXT"
+    assert sqlite_types.get(owner_id_str()) == "TEXT"
+    assert sqlite_types.get(acct_id_str()) == "TEXT"
+    assert sqlite_types.get(group_id_str()) == "TEXT"
+    assert sqlite_types.get(parent_road_str()) == "TEXT"
+    assert sqlite_types.get(label_str()) == "TEXT"
+    assert sqlite_types.get(road_str()) == "TEXT"
+    assert sqlite_types.get(base_str()) == "TEXT"
+    assert sqlite_types.get("need") == "TEXT"
+    assert sqlite_types.get("pick") == "TEXT"
+    assert sqlite_types.get(team_id_str()) == "TEXT"
+    assert sqlite_types.get(awardee_id_str()) == "TEXT"
+    assert sqlite_types.get(healer_id_str()) == "TEXT"
+    assert sqlite_types.get(time_id_str()) == "INTEGER"
+    assert sqlite_types.get(begin_str()) == "REAL"
+    assert sqlite_types.get(close_str()) == "REAL"
+    assert sqlite_types.get(addin_str()) == "REAL"
+    assert sqlite_types.get(numor_str()) == "REAL"
+    assert sqlite_types.get(denom_str()) == "REAL"
+    assert sqlite_types.get(morph_str()) == "INTEGER"
+    assert sqlite_types.get(gogo_want_str()) == "REAL"
+    assert sqlite_types.get(stop_want_str()) == "REAL"
+    assert sqlite_types.get(base_item_active_requisite_str()) == "TEXT"
+    assert sqlite_types.get(credit_belief_str()) == "REAL"
+    assert sqlite_types.get(debtit_belief_str()) == "REAL"
+    assert sqlite_types.get(credit_vote_str()) == "REAL"
+    assert sqlite_types.get(debtit_vote_str()) == "REAL"
+    assert sqlite_types.get(credor_respect_str()) == "REAL"
+    assert sqlite_types.get(debtor_respect_str()) == "REAL"
+    assert sqlite_types.get(fopen_str()) == "REAL"
+    assert sqlite_types.get(fnigh_str()) == "REAL"
+    assert sqlite_types.get("fund_pool") == "REAL"
+    assert sqlite_types.get(give_force_str()) == "REAL"
+    assert sqlite_types.get(mass_str()) == "REAL"
+    assert sqlite_types.get("max_tree_traverse") == "INTEGER"
+    assert sqlite_types.get("nigh") == "REAL"
+    assert sqlite_types.get("open") == "REAL"
+    assert sqlite_types.get("divisor") == "REAL"
+    assert sqlite_types.get(pledge_str()) == "INTEGER"
+    assert sqlite_types.get("problem_bool") == "INTEGER"
+    assert sqlite_types.get("purview_time_id") == "INTEGER"
+    assert sqlite_types.get(take_force_str()) == "REAL"
+    assert sqlite_types.get("tally") == "REAL"
+    assert sqlite_types.get(fund_coin_str()) == "REAL"
+    assert sqlite_types.get(penny_str()) == "REAL"
+    assert sqlite_types.get(respect_bit_str()) == "REAL"
+    assert sqlite_types.get(current_time_str()) == "INTEGER"
+    assert sqlite_types.get(amount_str()) == "REAL"
+    assert sqlite_types.get(month_label_str()) == "TEXT"
+    assert sqlite_types.get(hour_label_str()) == "TEXT"
+    assert sqlite_types.get(cumlative_minute_str()) == "INTEGER"
+    assert sqlite_types.get(cumlative_day_str()) == "INTEGER"
+    assert sqlite_types.get(weekday_label_str()) == "TEXT"
+    assert sqlite_types.get(weekday_order_str()) == "INTEGER"
+    assert sqlite_types.get(otx_road_delimiter_str()) == "TEXT"
+    assert sqlite_types.get(inx_road_delimiter_str()) == "TEXT"
+    assert sqlite_types.get(unknown_word_str()) == "TEXT"
+    assert sqlite_types.get(otx_word_str()) == "TEXT"
+    assert sqlite_types.get(inx_word_str()) == "TEXT"
+    assert sqlite_types.get(otx_label_str()) == "TEXT"
+    assert sqlite_types.get(inx_label_str()) == "TEXT"
+    assert sqlite_types.get(road_delimiter_str()) == "TEXT"
+    assert sqlite_types.get(c400_number_str()) == "INTEGER"
+    assert sqlite_types.get(yr1_jan1_offset_str()) == "INTEGER"
+    assert sqlite_types.get(quota_str()) == "REAL"
+    assert sqlite_types.get(monthday_distortion_str()) == "INTEGER"
+    assert sqlite_types.get(timeline_label_str()) == "TEXT"
 
 
 def test_get_allowed_curds_ReturnObj():
@@ -251,7 +473,8 @@ def _validate_brick_config(x_brick_config: dict):
         if fiscal_id_str() in sub_optional_args_keys:
             sub_optional_args_keys.remove(fiscal_id_str())
 
-        brick_optional_args_keys = set(brick_dict.get(optional_args_str()).keys())
+        brick_optional_args_dict = brick_dict.get(optional_args_str())
+        brick_optional_args_keys = set(brick_optional_args_dict.keys())
         # print(f"  {sub_optional_args_keys=}")
         # print(f"{brick_optional_args_keys=}")
         assert sub_optional_args_keys.issubset(brick_optional_args_keys)
@@ -262,6 +485,32 @@ def _validate_brick_config(x_brick_config: dict):
         if brick_dict.get(brick_type_str()) != filterunit_str():
             assert fiscal_id_str() in brick_required_args_keys
             assert time_id_str() in brick_required_args_keys
+
+        # sort_list = get_brick_elements_sort_order()
+        # x_count = 0
+        # sort_dict1 = {}
+        # sort_dict2 = {}
+        # for brick_arg in sort_list:
+        #     sort_dict1[x_count] = brick_arg
+        #     sort_dict1[brick_arg] = x_count
+
+        # for optional_arg in brick_optional_args_keys:
+        #     print(f"{optional_arg=} {brick_optional_args_dict=}")
+        #     optional_arg_dict = brick_optional_args_dict.get(optional_arg)
+        #     optional_arg_column_order = optional_arg_dict.get(column_order_str())
+        #     assert optional_arg_column_order != None
+        #     list_ref_arg = sort_list[optional_arg_column_order]
+        #     assert list_ref_arg != None
+        #     assert optional_arg == list_ref_arg
+
+        # # for required_arg in brick_required_args_keys:
+        #     print(f"{required_arg=} {brick_required_args_dict=}")
+        #     required_arg_dict = brick_required_args_dict.get(required_arg)
+        #     required_arg_column_order = required_arg_dict.get(column_order_str())
+        #     assert required_arg_column_order != None
+        #     list_ref_arg = sort_list[required_arg_column_order]
+        #     assert list_ref_arg != None
+        #     assert required_arg == list_ref_arg
 
 
 def test_get_brick_format_filenames_ReturnsObj():
